requirements.txt
cffi==1.7.0
colorlog>=2.7.0,<2.8
cryptography==1.4
Django[argon2]>=1.10,<1.11
django-cors-middleware>=1.3.1,<2.0.0
django-debug-toolbar==1.5
django-filter==0.13.0
django-rest-knox==2.2.0
django-rest-swagger>=2.0.4,<2.1.0
djangorestframework>=3.4.4,<3.5.0
idna==2.1
PyJWT>=1.4.0,<2.0.0
Markdown==2.6.6
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pyOpenSSL==16.0.0
six==1.10.0
sqlparse==0.2.0
